now patching ActiveRecord::Schema#define only without having to replace the entire class
